Added songs from 8th style
I added songs from beatmania IIDX 8th style. I also had to add a column that would include a Japanese written title (or a title with special letters or characters) as an alternative because I am still figuring out how R or SQL could recognize text written in a language that is not English. (This is completely experimental!)